+++ src/SMAPI/Framework/SCore.cs
@@ -161,7 +161,8 @@ namespace StardewModdingAPI.Framework
 
             // apply game patches
             new GamePatcher(this.Monitor).Apply(
-                new DialoguePatch(this.MonitorForGame, this.Reflection)
+                new DialogueErrorPatch(this.MonitorForGame, this.Reflection),
+                new LidgrenServerPatch()
             );
         }
 
@@ -208,7 +209,7 @@ namespace StardewModdingAPI.Framework
 
                 // override game
                 SGame.ConstructorHack = new SGameConstructorHack(this.Monitor, this.Reflection, this.Toolkit.JsonHelper);
-                this.GameInstance = new SGame(this.Monitor, this.MonitorForGame, this.Reflection, this.EventManager, this.InitialiseAfterGameStart, this.Dispose);
+                this.GameInstance = new SGame(this.Monitor, this.MonitorForGame, this.Reflection, this.EventManager, this.Toolkit.JsonHelper, this.ModRegistry, this.InitialiseAfterGameStart, this.Dispose, this.Settings.VerboseLogging);
                 StardewValley.Program.gamePtr = this.GameInstance;
 
                 // add exit handler
@@ -339,9 +340,6 @@ namespace StardewModdingAPI.Framework
         /// <summary>Initialise SMAPI and mods after the game starts.</summary>
         private void InitialiseAfterGameStart()
         {
-            // load settings
-            this.GameInstance.VerboseLogging = this.Settings.VerboseLogging;
-
             // load core components
             this.DeprecationManager = new DeprecationManager(this.Monitor, this.ModRegistry);
 
@@ -749,7 +747,7 @@ namespace StardewModdingAPI.Framework
             // log loaded content packs
             if (loadedContentPacks.Any())
             {
-                string GetModDisplayName(string id) => loadedMods.FirstOrDefault(p => id != null && id.Equals(p.Manifest?.UniqueID, StringComparison.InvariantCultureIgnoreCase))?.DisplayName;
+                string GetModDisplayName(string id) => loadedMods.FirstOrDefault(p => p.HasID(id))?.DisplayName;
 
                 this.Monitor.Log($"Loaded {loadedContentPacks.Length} content packs:", LogLevel.Info);
                 foreach (IModMetadata metadata in loadedContentPacks.OrderBy(p => p.DisplayName))
@@ -906,7 +904,7 @@ namespace StardewModdingAPI.Framework
                     if (this.ModRegistry.Get(dependency.UniqueID) == null)
                     {
                         string dependencyName = mods
-                            .FirstOrDefault(otherMod => otherMod.HasID() && dependency.UniqueID.Equals(otherMod.Manifest.UniqueID, StringComparison.InvariantCultureIgnoreCase))
+                            .FirstOrDefault(otherMod => otherMod.HasID(dependency.UniqueID))
                             ?.DisplayName ?? dependency.UniqueID;
                         errorReasonPhrase = $"it needs the '{dependencyName}' mod, which couldn't be loaded.";
                         return false;
@@ -970,7 +968,7 @@ namespace StardewModdingAPI.Framework
                     // get content packs
                     IContentPack[] contentPacks = this.ModRegistry
                         .GetAll(assemblyMods: false)
-                        .Where(p => p.IsContentPack && mod.Manifest.UniqueID.Equals(p.Manifest.ContentPackFor.UniqueID, StringComparison.InvariantCultureIgnoreCase))
+                        .Where(p => p.IsContentPack && mod.HasID(p.Manifest.ContentPackFor.UniqueID))
                         .Select(p => p.ContentPack)
                         .ToArray();